chatbotroutes_service/requirements.txt
fastapi
pydantic
requests
python-dotenv
langdetect
gunicorn
uvicorn